+++ src/Contrib/Grpc/GrpcTransportFactory.php
@@ -84,7 +84,7 @@ final class GrpcTransportFactory implements TransportFactoryInterface
             $opts,
             $method,
             $headers,
-            (int) ($timeout * self::MILLIS_PER_SECOND),
+            (int) ($timeout * (float) self::MILLIS_PER_SECOND),
         );
     }
